+++ app/controllers/staff/organisations_controller.rb
@@ -13,7 +13,7 @@ class Staff::OrganisationsController < Staff::BaseController
     @organisation_presenter = OrganisationPresenter.new(organisation)
 
     project_activities = FindProjectActivities.new(organisation: organisation, user: current_user).call(eager_load_parent: false)
-    third_party_project_activities = FindThirdPartyProjectActivities.new(organisation: organisation, user: current_user).call
+    third_party_project_activities = FindThirdPartyProjectActivities.new(organisation: organisation, user: current_user).call(eager_load_parent: false)
 
     respond_to do |format|
       format.html do